requirements.txt
google-genai>=1.0.0
Pillow==12.1.0
pillow-heif==1.3.0
rich==14.0.0
python-dotenv==1.1.0
tenacity==9.1.2
flask>=3.0.0
